requirements.txt
# BP 神经网络项目依赖
# 安装命令：pip install -r requirements.txt

# 深度学习框架
torch>=2.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 机器学习工具
scikit-learn>=1.3.0

# 超参数优化
optuna>=3.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置与序列化
pyyaml>=6.0
joblib>=1.3.0

# 进度条
tqdm>=4.65.0
